A new gem that captures mock_model and stub_model.
This is a first pass extraction of the mock_model and
stub_model code. It doesn't have a lot of the supporting
configuration in it that other rspec projects currently
have and has not yet been stripped down to remove things
that are not necessary. I'm not happy with the need to
replicate the generation of a sample rails project but
it was the quickest way to get this off the ground for
now.

For rspec/rspec-rails#894