Fix === for multiple mocks of the same model
When the same model is mocked more than once in the same context,
comparing with the model class using `#===` would only return true for
the first of the mocks. All following mocks would return false when
compared to the original model class using `#===`.

For `#===` to work correctly, we register each mock in a "mock_store"
on the `RSpec::ActiveModel::Mocks::Mocks` module. The store is
implemented as an RSpec stub, so it is reset after each spec example.

Fixes #60